Move all implicits to implicits package
