+++ src/main/scala/libra/implicits/AmountImplicits.scala
@@ -1,0 +1,12 @@
+package libra
+package implicits
+
+import ops.base.Show
+
+trait AmountImplicits {
+
+  implicit def amountShow: Show[Amount] = Show[Amount]("N")
+
+  implicit def moleShow: Show[Mole] = Show[Mole]("mol")
+
+}
