+++ src/main/scala/libra/implicits/CurrentImplicits.scala
@@ -1,7 +1,17 @@
 package libra
 package implicits
 
+import units.MetricUnit
+import ops.base.Show
+
+import singleton.ops._
+
 trait CurrentImplicits {
+
+  implicit def currentShow: Show[Current] = Show[Current]("I")
+
+  implicit def metricCurrentShow[I <: XInt](implicit s: Show[I]): Show[MetricUnit[I, Current]] = Show(s"${s()}A")
+
   implicit final class CurrentOps[A](val a: A) {
     def kA: QuantityOf[A, Current, Kiloampere] = Quantity(a)
     def A: QuantityOf[A, Current, Ampere] = Quantity(a)
