+++ src/main/scala/libra/units/MetricUnit.scala
@@ -1,0 +1,6 @@
+package libra
+package units
+
+import singleton.ops._
+
+trait MetricUnit[I <: XInt, D] extends UnitOfMeasure[D]
