+++ src/main/scala/libra/units/TimeUnits.scala
@@ -1,11 +1,8 @@
 package libra
 package units
 
-import ops.base.{Show, ConversionFactor}
-import spire.math._
-import singleton.ops._
 
-trait TimeUnits extends MetricUnitImplicits {
+trait TimeUnits {
 
   type Time
 
@@ -23,16 +20,4 @@ trait TimeUnits extends MetricUnitImplicits {
   trait Day extends UnitOfMeasure[Time]
   trait Hour extends UnitOfMeasure[Time]
 
-  implicit def timeShow: Show[Time] = Show[Time]("T")
-
-  implicit def metricTimeShow[I <: XInt](implicit s: Show[I]): Show[MetricUnit[I, Time]] = Show(s"${s()}s")
-
-
-  implicit def dayShow: Show[Day] = Show[Day]("days")
-  implicit def hourShow: Show[Hour] = Show[Hour]("hours")
-
-  implicit def dayHourConversion[A](
-    implicit c: ConvertableTo[A]
-  ): ConversionFactor[A, Time, Day, Hour] =
-    new ConversionFactor(c.fromInt(24))
 }
